connect: carve fleet.rs (pure move)
The browser-synced user records leave main.rs: signed fleet-target
sync/forget with sanitization and owned-daemon canonicalization, the
daemon label/revoke endpoints, and the end-to-end encrypted credential
vault blob (publish/fetch with its revision and MAC-presence
ratchets), plus the fleet and vault tests and their two blob helpers.
